Added Crimson Hearts (naturally generate), Added a chair block that spawns in NPCs.
The Wooden Chair block spawns in NPCs using scoreboards.

The nurse will spawn if the life crystal scoreboard is 1 or more

The merchant will spawn if the coin scoreboard is more then 30. The coin scoreboard will increase when killing a green slime (will include more enemies in the next build)

The Demolitionist will spawn if the coins scoreboard is more then 30 and if the explosives scoreboard is more then 30. The explosives scoreboard will increase if surface chest are broken (will include other chest too in the next build).

The arms dealer will spawn if the ammo scoreboard is more then one. The ammo scoreboard can be increased if crimson hearts are broken. (will include the corruption version in the next build).