2.7.0 release candidate 3
Adjusted the distance fog renders from the player.
Cobwebs now slow the player when try to move though them.
Updated the Hellforge's model and texture.
Adjusted the Antlion's model to make it look more like it's in the ground.
Fixed a bug where players could attack NPCs.
Fixed the bottom texture of the Cursed Skull.
Removed Rocket Boots, and the Cellphone.
Changed rarity of biomes.
Readded the corruption Thorns.
The corruption Thorns now slow down the player and damage them even if the block is naturally placed instead of only when it is placed by a player.
Crimson Hearts and Crimson Altars now spawn in the bottom of chasms and underground like Shadow Orbs and Demon Alters do.
Fixed the sun's texture.
Fixed a bug where the Moonglow didn't drop it's self.
Fixed a bug where the Crimson Rod didn't require mana to use.
Antlion Chargers now spawn in the world.
Fixed a bug where players could open a container when interacting with an other player.
Fixed a bug where the bottom of crimson, corruption, and mushroom grasses would have an outdated texture.
Fixed a bug where the Spiky Ball was called the "Flower of Fire".
Fixed a bug where the Demon Scythe was called the "Demon Scyten" in-game.
Fixed a bug where the Magic Mirror would have a strange font and incorrect colors.
Fixed a bug where the Cream Soda and Minishark items tooltips where on the wrong line.
Fixed a bug where the Gold Chest would say "****" in chat when placed.
Fixed a bug where the Mining Potion's tooltip didn't end with percent.